% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchyLong.R
\name{hierarchyLong}
\alias{hierarchyLong}
\title{hierarchyLong}
\usage{
hierarchyLong(ee, supv)
}
\arguments{
\item{ee}{A list of values representing employees (e.g. employee IDs).}

\item{supv}{A list of values representing the supervisors of the
employees. These values should be of the same data type as the employee values.}
}
\value{
data frame
}
\description{
The hierarchyLong function transforms a standard set of unique employee and
supervisor identifiers into a long format that can be used to aggregate employee data
by a particular line of leadership (i.e. include everyone who rolls up to Susan). The
function returns a long data frame consisting of one row per employee for every supervisor
above them, up to the top of the tree (i.e. the CEO in your organization). The levels represent
the number of supervisors from the employee (starting with "1" for an employee's direct supervisor).
}
\examples{
ee = c("Dale","Bob","Jill","Mark","Julie","Andrea","Susan")
supv = c("Julie","Julie","Julie","Andrea","Susan","Susan","George")
hierarchyLong(ee,supv)
}
