% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchyStats.R
\name{hierarchyStats}
\alias{hierarchyStats}
\title{hierarchyStats}
\usage{
hierarchyStats(ee, supv)
}
\arguments{
\item{ee}{A vector containing unique identifiers for employees.}

\item{supv}{A vector containing unique identifiers for supervisors. These values should be
of the same type as the employee values.}
}
\value{
list
}
\description{
The hierarchyStats function computes summary statistics and span of control metrics
from a standard set of unique employee and supervisor identifiers (employee IDs, email addresses, etc.).
}
\examples{
ee = c("Dale@hR.com","Bob@hR.com","Julie@hR.com","Andrea@hR.com")
supv = c("Julie@hR.com","Julie@hR.com","Andrea@hR.com","Susan@hR.com")
hierarchyStats(ee,supv)
}
