% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchyValid.R
\name{hierarchyValid}
\alias{hierarchyValid}
\title{hierarchyValid}
\usage{
hierarchyValid(ee, supv)
}
\arguments{
\item{ee}{A vector containing unique identifiers for employees.}

\item{supv}{A vector containing unique identifiers for supervisors. These values should be
of the same type as the employee values.}
}
\value{
logical
}
\description{
The hierarchyValid function considers a standard set of unique employee and 
supervisor identifiers (employee IDs, email addresses, etc.) and validates the completeness
and quality of the two input vectors representing the overall hierarchy.
}
\examples{
ee = c("Dale@hR.com","Bob@hR.com","Julie@hR.com","Andrea@hR.com")
supv = c("Julie@hR.com","Julie@hR.com","Andrea@hR.com","Susan@hR.com")
hierarchyValid(ee,supv)
}
