% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR-package.R
\docType{package}
\name{HR-package}
\alias{HR}
\alias{HR-package}
\title{HR: Iq Test Data}
\description{Aggregate IQ test data by time and the passing mark.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Suvo198/HR/blob/main/HR.csv}
  \item Report bugs at \url{https://github.com/Suvo198/HR/issues}
}

}
\author{
\strong{Maintainer}: Subhadeep Banik Horst \email{baniksuvo6991@hotmail.com} (\href{https://orcid.org/0009-0007-2216-9340}{ORCID})

Authors:
\itemize{
  \item Alison Hill \email{baniksuvo6991@hotmail.com} (\href{https://orcid.org/0009-0007-2216-9340}{ORCID})
}

}
\keyword{internal}
