% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\docType{data}
\name{HR}
\alias{HR}
\title{Sample for the A/B test for the learning growth}
\format{
A tibble with 344 rows and 8 variables:
\describe{
\item{Time}{The total time is 10 minutes for an interview. )}
\item{Marks}{(Total marks is the 100.)}
}
}
\usage{
HR
}
\description{
Includes measurements for penguin species, island in Palmer Archipelago,
size (Aggregate data by time and the passing mark.\code{\link{HR})}.
}
\keyword{datasets}
