% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-to-file.R
\name{path_to_file}
\alias{path_to_file}
\title{Get file path to \code{HR.csv} files}
\source{
This function is adapted from \code{readxl::readxl_example()}.
}
\usage{
path_to_file(path = NULL)
}
\arguments{
\item{path}{Name of file in quotes with extension;
\code{"HR.csv"} will work.
If \code{NULL}, the example files will be listed.}
}
\description{
HR comes with the csv file in its \code{inst/extdata}
directory. This function make them easy to access.
}
\details{
The data in each file is the same as \code{\link{HR}},
but without the variable types.
}
\examples{
path_to_file("HR.csv")
head(read.csv(path_to_file("HR.csv")))
}
