% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workforcePlan.R
\name{workforcePlan}
\alias{workforcePlan}
\title{workforcePlan}
\usage{
workforcePlan()
}
\description{
Launch a simple, interactive workforce planning worksheet that helps managers and
team leaders to execute basic workforce planning tasks and plan ahead for hiring, turnover, and other
factors that influence a team's talent structure. Data analysts can use this alongside
team leaders to convey change and proactively think about recruitment, etc.
}
