% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchyLong.R
\name{hierarchyLong}
\alias{hierarchyLong}
\title{hierarchyLong}
\usage{
hierarchyLong(ee, supv)
}
\arguments{
\item{ee}{A list of values representing employees (e.g. employee IDs).}

\item{supv}{A list of values representing the supervisors of the 
employees. These values should be of the same data type as the employee values.}
}
\value{
data frame
}
\description{
This function transforms a standard set of 
employee and supervisor identifiers into a long format representing
full hierarchies. This is especially useful for narrowing in on a 
subset of the workforce that reports up to a single person. The 
function returns a long data frame consisting of one row per employee 
for every supervisor above them, up to the CEO. The levels represent 
the number of supervisors from the employee (starting with "1" for 
an employee's direct supervisor).
}
\examples{
ee = c("Dale","Bob","Jill","Mark","Julie","Andrea","Susan")
supv = c("Julie","Julie","Julie","Andrea","Susan","Susan","George")
hierarchyLong(ee,supv)
}
