% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchyWide.R
\name{hierarchyWide}
\alias{hierarchyWide}
\title{hierarchyWide}
\usage{
hierarchyWide(ee, supv)
}
\arguments{
\item{ee}{A list of values representing employees (e.g. employee IDs).}

\item{supv}{A list of values representing the supervisors of the employees. These values should be
of the same type as the employee values.}
}
\value{
data frame
}
\description{
This function transforms a standard set of employee 
and supervisor identifiers into a wide format with complete hierarchies. 
The resulting data can be aggregated by a particular person (i.e. include 
everyone who rolls up to "Julie") to include all direct and indirect reports.
}
\examples{
ee = c("Dale","Bob","Jill","Mark","Julie","Andrea","Susan")
supv = c("Julie","Julie","Julie","Andrea","Susan","Susan","George")
hierarchyWide(ee,supv)
}
