% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchyWide.R
\name{hierarchyWide}
\alias{hierarchyWide}
\title{hierarchyWide}
\usage{
hierarchyWide(ee, supv)
}
\arguments{
\item{ee}{A value representing an employee}

\item{supv}{A value representing the employee's supervisor}
}
\description{
This function takes two lists representing employee and supervisor
identifiers (name, ID, etc.) and returns a wide dataframe consisting of
a single row per employee and their reporting hierarchy in a wide format.
}
