% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchyWide.R
\name{hierarchyWide}
\alias{hierarchyWide}
\title{hierarchyWide}
\usage{
hierarchyWide(ee, supv)
}
\arguments{
\item{ee}{A list of values representing employees (e.g. employee IDs).}

\item{supv}{A list of values representing the supervisors of the employees. These values should be
of the same type as the employee values.}
}
\value{
data frame
}
\description{
The hierarchyWide function transforms a standard set of unique employee and supervisor
identifiers into a wide format that can be used to aggregate employee data by a particular line
of leadership (i.e. include everyone who rolls up to Susan). The function returns a wide data frame
with an additional column for every level in the hierarchy, starting from the top of the
tree (i.e. "Supv1" is the CEO in your organization).
}
\examples{
ee = c("Dale","Bob","Jill","Mark","Julie","Andrea","Susan")
supv = c("Julie","Julie","Julie","Andrea","Susan","Susan","George")
hierarchyWide(ee,supv)
}
