% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchyLong.R, R/hierarchyWide.R
\name{hierarchyWide}
\alias{hierarchyWide}
\alias{hierarchyWide}
\title{hierarchyLong}
\usage{
hierarchyWide(ee, supv)

hierarchyWide(ee, supv)
}
\arguments{
\item{ee}{A list of values representing employees (e.g. employee IDs).}

\item{supv}{A list of values representing the supervisors of the employees. These values should be
of the same type as the employee values.}

\item{ee}{A list of values representing employees (e.g. employee IDs).}

\item{supv}{A list of values representing the supervisors of the employees. These values should be
of the same type as the employee values.}
}
\value{
data frame

data frame
}
\description{
This function takes employee and supervisor
identifiers (name, ID, etc.) and returns a long data frame consisting of
one row per employee for every supervisor above them, up to the CEO.

The resulting format is useful for quickly filtering to a specific
part of the workforce that directly and indirectly reports up through
a particular leader.

This function takes employee and supervisor
identifiers (name, ID, etc.) and returns a wide data frame consisting of
a single row per employee and their respective reporting hierarchy in a wide format.

The resulting format is very useful for subsequent aggregation of employee data
for a particular leadership tree.
}
\examples{
ee = c("Dale","Bob","Julie","Susan")
supv = c("Julie","Julie","Susan","George")
hierarchyLong(ee,supv)
ee = c("Dale","Bob","Julie","Susan")
supv = c("Julie","Julie","Susan","George")
hierarchyWide(ee,supv)
}
