% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idPad.R
\name{idPad}
\alias{idPad}
\title{idPad}
\usage{
idPad(id, len)
}
\arguments{
\item{id}{An employee ID.}

\item{len}{The standard length requirement for each employee ID. Leading zeroes
will be added to fill up the space to meet the requirement.}
}
\value{
character value
}
\description{
This function converts an employee ID into a specified format
that includes leading zeroes. HR systems often create employee IDs that have leading zeroes.
The leading zeroes can be lost when they are worked on in Excel or other tools. It's usually
helpful to convert the IDs back the correct format when performing lookups.
}
\examples{
# Using a string input
id = "12345"
len = 9
idPad(id,len)

"000012345"

# Using a numeric input
id = 54321
len = 9
idPad(id,len)

"000054321"
}
