% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visRoleChanges.R
\name{visRoleChanges}
\alias{visRoleChanges}
\title{visRoleChanges}
\usage{
visRoleChanges(df)
}
\arguments{
\item{df}{A data frame with two columns. Column 1 should be the "before" job titles. Column 2 should be the "after" job titles".}

\item{supv}{A list of values representing the supervisors of the employees. These values should be
of the same type as the employee values.}
}
\value{
visNetwork
}
\description{
This function requires one data frame with two columns representing
the titles of employees from two different time periods.This function is useful
for visualizing "before-and-after" role changes.
}
\examples{
df = data.frame(from=c(rep("builder",10),"tester","tester","builder","manager","builder","recruiter"),
               to=c("tester",rep("builder",10),"tester","builder","manager","builder","builder"))
visRoleChanges(df)
}
